+++ app/frontend/src/hooks/index.ts
@@ -1,3 +1,4 @@
+export * from './useChatting';
 export * from './useDebounce';
 export * from './useMap';
 export * from './useModal';
